Ignores calls coming from non focus user.
Adds a property to control the resource used by focus, default is 'focus'.
Cleans listener from DTMF operation set.
Ignores calls coming from non focus user, when p2p is enabled we receive session-initiate from the other participant in the room if he is alone.
Updates fmj with a fix for a NPE calling twice bye on rtcp handler.
Updates protocol-media with a fix to close audio translator.
Updates protocol-jabber bundle with correctly parsing jingleIQ, group-bundle.
Updates libjitsi that adds device AudioSilenceCaptureDevice:noTransferData, which doesn't start a thread for transferring data, which is not needed in case of translator mode.