Waits for reply from jicofo on invite.
If we do not wait we can end up joining before jicofo and this breaks
jicofo logic. The side effect is that jicofo does not detect that there
is someone in the room and leaves it after 15 seconds if it is only
jigasi in there and the call is dropped for jigasi.