Detects incoming media dropped on sip side.
If we detect dropped media, print messages, add stats and when it
resumes stat is updated and message is printed that RTP resumed.
The setting is a threshold of time after which we consider media
for dropped, -1 disables the behaviour, default is 10 seconds and
enabled. There is also an option that, when enabled, calls can be
dropped when media stops.