ref(google transcriptions): handle only the first result
The first result contains the most stable part of the transcription and
the second one contains unstable, work in progress part that will
eventually stabilize and make it to the first result(the stable part).

This reduces the number of messages sent to the MUC and the overall
jitterness seen on the client.