Fix problem loading smack classes.
Seems that https://github
.com/jitsi/jigasi/pull/258/commits
/dfd269d33cb775b036649e6ef5cca9bf5ebd82d3 is not resolving the problem
and we still experience it.

While doing XMPPTCPConnection.setUseStreamManagementDefault we start
loading SmackInitialization class which waits for ProviderManager class
to be loaded but ProviderManager is waiting to load SmackInitialization.
Moving XMPPTCPConnection.setUseStreamManagementDefault in the main
thread will ensure we load SmackInitialization earlier and should fix
the problem.