feat: Exposes the "stress_level" metric.
This commit exposes a new metric (called "stress_level") that can be
used to determined the load factor of a Jigasi instance. It is currently
based on the number of concurrent conferences with a default normalizing
constant of 100. This constant needs to be fine tuned based on the
specific instance type that Jigasi is running and it needs to represent
the max number of conferences that Jigasi can handle correctly on the
specific instance type (i.e. if Jigasi has more conferences it is
overloaded and it may experience abnormal operation).