fix: Fixes dynamic adding shards.
There can be concurrency issue where we create and load the account and at the time the provider is registered as osgi service we got notified and we try to load the password through the factory (which uses the configuration storage service) the password is still not set and written by the config service. Loading the password directly from the account properties.