feat: Implements a simple message to be set in call to check for server availability.
If there is a short network disruption and sip connection reconnects on TCP we may never receive the sip call BYE so we never cleanup and leak calls.
This is a simple heartbeat like, sending a sip INFO message and expecting a reply, if two consecutive miss a response we drop the call, anyway it was dropped at server-side.