fix(whisper): Fixes leaking whisper connections.
Moves setting transcriber room name earlier, so we can use it in creating the transcriber sessions.

When a participant never talked, is missing in one of the participants lists and those got desync and we never remove the websocket from the pool, which results in reusing it for the same room (the socket has timeouted and disconnected) and there is a NPE that prevents transcriptions in the meeting.
2023-11-30 13:09:17.645 SEVERE: [188854] org.jivesoftware.smack.AbstractXMPPConnection.lambda$invokeStanzaCollectorsAndNotifyRecvListeners$8: Exception in packet listener
java.lang.NullPointerException
	at org.jitsi.jigasi.transcription.WhisperWebsocket.addParticipantIfNotExists(WhisperWebsocket.java:362)
	at org.jitsi.jigasi.transcription.WhisperWebsocket.addListener(WhisperWebsocket.java:373)
	at org.jitsi.jigasi.transcription.WhisperTranscriptionService$WhisperWebsocketStreamingSession.addTranscriptionListener(WhisperTranscriptionService.java:170)
	at org.jitsi.jigasi.transcription.Participant.joined(Participant.java:524)